Clean up organization of table related types.
