+++ ui/apps/platform/src/types/table.ts
@@ -1,0 +1,22 @@
+import { ReactElement } from 'react';
+
+export type { ThProps } from '@patternfly/react-table';
+
+export type SortDirection = 'asc' | 'desc';
+
+export type SortOption = {
+    field: string;
+    direction: SortDirection;
+};
+
+export type TableColumn = {
+    Header: string;
+    accessor: string;
+    Cell?: ({ original, value }) => ReactElement | string;
+    sortField?: string;
+};
+
+export type TableSortOption = {
+    field: string;
+    reversed: boolean;
+};
